refact: improve project breadcrunbs component
